+++ packages/nextra-theme-docs/src/docs-sidebar.js
@@ -1,0 +1,165 @@
+import React, { useState, useEffect, useContext, createContext } from 'react'
+import Slugger from 'github-slugger'
+import { useRouter } from 'next/router'
+import Link from 'next/link'
+
+import { useActiveAnchor } from './misc/active-anchor'
+import { getFSRoute } from './utils/get-fs-route'
+
+const TreeState = new Map()
+const MenuContext = createContext(false)
+
+function Folder({ item, anchors }) {
+  const { asPath, locale } = useRouter()
+  const routeOriginal = getFSRoute(asPath, locale)
+  const route = routeOriginal.split('#')[0] + '/'
+  const active = route === item.route + '/'
+  const { defaultMenuCollapsed } = useContext(MenuContext)
+  const open = TreeState[item.route] ?? !defaultMenuCollapsed
+  const [_, render] = useState(false)
+
+  useEffect(() => {
+    if (active) {
+      TreeState[item.route] = true
+    }
+  }, [active])
+
+  return (
+    <li className={open ? 'active' : ''}>
+      <button
+        onClick={() => {
+          if (active) return
+          TreeState[item.route] = !open
+          render(x => !x)
+        }}
+      >
+        {item.title}
+      </button>
+      <div
+        style={{
+          display: open ? 'initial' : 'none'
+        }}
+      >
+        <Menu dir={item.children} base={item.route} anchors={anchors} />
+      </div>
+    </li>
+  )
+}
+
+function File({ item, anchors }) {
+  const { setMenu } = useContext(MenuContext)
+  const { asPath, locale } = useRouter()
+  const route = getFSRoute(asPath, locale) + '/'
+  const active = route === item.route + '/'
+  const slugger = new Slugger()
+  const activeAnchor = useActiveAnchor()
+
+  const title = item.title
+  // if (item.title.startsWith('> ')) {
+  // title = title.substr(2)
+  if (anchors && anchors.length) {
+    if (active) {
+      let activeIndex = 0
+      const anchorInfo = anchors.map((anchor, i) => {
+        const text = innerText(anchor) || ''
+        const slug = slugger.slug(text)
+        if (activeAnchor[slug]) {
+          activeIndex = i
+        }
+        return { text, slug }
+      })
+
+      return (
+        <li className={active ? 'active' : ''}>
+          <Link href={item.route}>
+            <a>{title}</a>
+          </Link>
+          <ul>
+            {anchors.map((_, i) => {
+              const { slug, text } = anchorInfo[i]
+              const isActive = i === activeIndex
+
+              return (
+                <li key={`a-${slug}`}>
+                  <a
+                    href={'#' + slug}
+                    onClick={() => setMenu(false)}
+                    className={isActive ? 'active-anchor' : ''}
+                  >
+                    <span className="flex text-sm">
+                      <span className="opacity-25">#</span>
+                      <span className="mr-2"></span>
+                      <span className="inline-block">{text}</span>
+                    </span>
+                  </a>
+                </li>
+              )
+            })}
+          </ul>
+        </li>
+      )
+    }
+  }
+
+  return (
+    <li className={active ? 'active' : ''}>
+      <Link href={item.route}>
+        <a onClick={() => setMenu(false)}>{title}</a>
+      </Link>
+    </li>
+  )
+}
+
+function Menu({ dir, anchors }) {
+  return (
+    <ul>
+      {dir.map(item => {
+        if (item.children) {
+          return <Folder key={item.name} item={item} anchors={anchors} />
+        }
+        return <File key={item.name} item={item} anchors={anchors} />
+      })}
+    </ul>
+  )
+}
+
+function Sidebar({ show, directories, anchors }) {
+  return (
+    <aside
+      className={`h-screen bg-white dark:bg-dark flex-shrink-0 w-full md:w-64 md:block fixed md:sticky z-10 ${
+        show ? '' : 'hidden'
+      }`}
+      style={{
+        top: '4rem',
+        height: 'calc(100vh - 4rem)'
+      }}
+    >
+      <div className="sidebar border-gray-200 dark:border-gray-900 w-full p-4 pb-40 md:pb-16 h-full overflow-y-auto">
+        <Menu dir={directories} anchors={anchors} />
+      </div>
+    </aside>
+  )
+}
+
+export default function DocsSidebar({ directories, anchors, config }) {
+  const [menu, setMenu] = useState(false)
+
+  useEffect(() => {
+    if (menu) {
+      document.body.classList.add('overflow-hidden')
+    } else {
+      document.body.classList.remove('overflow-hidden')
+    }
+  }, [menu])
+
+  return (
+    <MenuContext.Provider
+      value={{
+        setMenu,
+        defaultMenuCollapsed: !!config.defaultMenuCollapsed
+      }}
+    >
+      <Sidebar show={menu} anchors={anchors} directories={directories} />
+    </MenuContext.Provider>
+  )
+}
